requirements.txt

MySQL-python==1.2.5
numpy==1.13.3
pandas==0.20.1
pip==9.0.1
requests==2.14.2
virtualenv==15.1.0
Pillow==5.0.0
